Make compilers use variable sized integers.
Also, update to the newest DelftProto VM, that always uses variable sized integers.

git-svn-id: file:///Users/jakebeal/projects/_fin/proto/svn/proto/trunk@557 4c76e1c1-4b85-4c65-98a1-c45dc8243d55

+++ proto/src/compiler/compiler.h
@@ -199,7 +199,7 @@ class ProtoKernelEmitter : public CodeEmitter {
   Instruction *branch_primitive_instruction(OperatorInstance *oi);
   Instruction *literal_to_instruction(ProtoType *l, OperatorInstance *context);
   Instruction *scalar_literal_instruction(ProtoScalar *scalar);
-  Instruction *integer_literal_instruction(uint16_t value);
+  Instruction *integer_literal_instruction(unsigned int value);
   Instruction *float_literal_instruction(float value);
   Instruction *tuple_literal_instruction(ProtoTuple *tuple,
       OperatorInstance *context);
